move app from repo test to e2e section
